fix recent_modders_list is not persisted
